Rename nmigen.lib.roundrobin to nmigen.lib.scheduler
