+++ nmigen/lib/scheduler.py
